new transaction route

+++ api/src/controllers/transactionController.js
@@ -1,6 +1,6 @@
 const Transaction = require('../models/transaction');
 const Budget = require('../models/budget');
-const { fetchCategoryFromAI } = require('../utils/functions');
+const { fetchCategoryFromAI, getDataWithCaching } = require('../utils/functions');
 const redisClient = require('../config/redis');
 const { GoogleGenerativeAI } = require("@google/generative-ai");
 const { mainServerLogger } = require('../utils/logger');
@@ -16,6 +16,20 @@ async function routeWrapper(req, res, next, handler) {
         next(err)
     }
 }
+
+const getTransaction = async (req, res, next) => {
+    await routeWrapper(req, res, next, async () => {
+        const transaction = await getDataWithCaching(redisClient, `transaction:${req.user.uid}:${req.params.transaction_id}`, async () => {
+            return await Transaction.findOne({ _id: req.params.transaction_id, user_id: req.user.uid });
+        })
+        if (!transaction) {
+            const error = new Error("Forbidden")
+            error.status = 403
+            return next(error)
+        }
+        res.json({ transaction });
+    })
+};
 
 const createTransaction = async (req, res, next) => {
     await routeWrapper(req, res, next, async () => {
@@ -63,6 +77,7 @@ const updateTransaction = async (req, res, next) => {
                 req.body.budget_id, { $addToSet: { categories: req.body.category } }
             );
         }
+        await redisClient.del(`transaction:${req.user.uid}:${req.params.transaction_id}`)
         await redisClient.del(`transactions:${req.user.uid}:${req.body.budget_id}`)
             
         res.json('OK')
@@ -71,7 +86,7 @@ const updateTransaction = async (req, res, next) => {
 
 const deleteTransaction = async (req, res, next) => {
     await routeWrapper(req, res, next, async () => {
-        const { budget_id, amount } = await Transaction.findOneAndDelete({ _id: req.params.transaction_id }).select('budget_id amount')
+        const { budget_id, amount } = await Transaction.findOneAndDelete({ _id: req.params.transaction_id, user_id: req.user.uid }).select('budget_id amount')
         await Budget.findByIdAndUpdate(
             budget_id,
             {
@@ -79,12 +94,14 @@ const deleteTransaction = async (req, res, next) => {
                 $inc: { current_balance: -amount }
             }
         );
-        await redisClient.del(`transactions:${req.user.uid}:${budget_id}`)
+        await redisClient.del(`transaction:${req.user.uid}:${req.params.transaction_id}`)
+        await redisClient.del(`transactions:${req.user.uid}:${req.body.budget_id}`)
         res.json('OK')
     })
 };
 
 module.exports = { 
+    getTransaction,
     createTransaction,
     updateTransaction,
     deleteTransaction,
